requirements.txt
Django==1.10.2
django-cors-headers==2.2.0
django-tinymce4-lite==1.7.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Pillow==3.4.1
PyJWT==1.6.1
pytz==2017.2
